hasMemoメソッドで事前に表示内容を決定するのではなく、Memoクラスの振る舞いの結果で表示内容を決定するよう修正

+++ 04.class/lib/memo.js
@@ -4,11 +4,6 @@ import { SelectableList } from "./selectable-list.js";
 export class Memo {
   #memoDB = new MemoDB();
 
-  async hasMemo() {
-    const memos = await this.#memoDB.selectAll();
-    return memos.length !== 0;
-  }
-
   async add(lines) {
     await this.#memoDB.insert(lines);
     return await this.#memoDB.close();
@@ -23,20 +18,32 @@ export class Memo {
 
   async #selectFromList(instruction) {
     const memos = await this.#memoDB.selectAll();
-    const list = new SelectableList(memos, instruction);
-    return await list.selectMemo();
+    if (memos.length === 0) {
+      return null;
+    } else {
+      const list = new SelectableList(memos, instruction);
+      return await list.selectMemo();
+    }
   }
 
   async full() {
     const memoID = await this.#selectFromList("see");
-    const selectedMemo = await this.#memoDB.selectByID(memoID);
-    await this.#memoDB.close();
-    return selectedMemo.body;
+    if (memoID === null) {
+      return null;
+    } else {
+      const selectedMemo = await this.#memoDB.selectByID(memoID);
+      await this.#memoDB.close();
+      return selectedMemo.body;
+    }
   }
 
   async delete() {
     const memoID = await this.#selectFromList("delete");
-    await this.#memoDB.delete(memoID);
-    return await this.#memoDB.close();
+    if (memoID === null) {
+      return null;
+    } else {
+      await this.#memoDB.delete(memoID);
+      await this.#memoDB.close();
+    }
   }
 }
